requirements.txt
asgiref==3.7.2
confluent-kafka==2.3.0
Django==5.0
djangorestframework==3.14.0
pika==1.3.2
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
